python-catch-up/pyproject.toml
[tool.poetry]
name = "python-catch-up"
version = "0.1.0"
description = ""
authors = ["imakyo97 <kyoheishogun.1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
tortoise-orm = "^0.20.0"
fastapi-admin = "^1.0.4"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
trio = "^0.24.0"
asgi-lifespan = "^2.1.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-mock = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "python_catch_up"
testpaths = ["tests",]
